test: 테스트 기능 추가 및 목록 나눔

+++ src/main/kotlin/racingcar/Input.kt
@@ -4,5 +4,5 @@ import camp.nextstep.edu.missionutils.Console
 
 class Input {
     fun carNames() = Console.readLine().split(",")
-    fun moveCount() = Console.readLine().toInt()
+    fun moveCount() = Console.readLine()
 }
